reverse_funnel_outreach/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reverse-funnel-outreach"
version = "0.1.0"
description = "Reverse funnel outreach: harvest emails (Crawl4AI + Scrapy + Trafilatura), ConduitScore scans, cold-email rendering"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "jinja2>=3.1.0",
  "typer>=0.12.0",
  "rich>=13.0.0",
  "trafilatura>=1.8.0",
  "scrapy>=2.11.0",
  "crawl4ai>=0.4.0",
  "dnspython>=2.6.1",
  "extruct>=0.16.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.4.0"]

[project.scripts]
rf-outreach = "reverse_funnel_outreach.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/reverse_funnel_outreach"]

[tool.hatch.build.targets.sdist]
include = ["src/reverse_funnel_outreach", "README.md"]
